qbittorrentFixStalled/requirements.txt
python-dotenv==1.0.0
qbittorrent-api==2023.6.50 
scheduler==0.8.4